Updated project files
Automation now runs all the way through. Currently updating it to incorporate Orchestrator queues and assets as well as tidying it up along the way. Adding in validations, exception handling and any debugging as I'm tidying it up. Need to add in reporting. If I add in trigger for the cut-off point, I may split it into 2 processes, one for dealing with the user requests before cutoff point and the other for dealing with the content aggregation and delivery.